// src/app/store/book/book.types.ts
export enum BookActionsType {
  LOAD_BOOK = '[Book Page]Load Book',
  LOAD_BOOK_SUCCESS = '[Book Page]Load Book Success',
  LOAD_BOOK_FAIL = '[Book Page]Load Book Fail',
  ADD_BOOK = '[Book Page]Add Book',
  ADD_BOOK_SUCCESS = '[Book Page]Add Book Success',
  UPDATE_BOOK = '[Book Page]Update Book',
  UPDATE_BOOK_SUCCESS = '[Book Page]Update Book Success',
  DELETE_BOOK = '[Book Page]Delete Book',
  DELETE_BOOK_SUCCESS = '[Book Page]Delete Book Success',
  GET_BOOK = '[Book Page]Get Book',
  GET_BOOK_SUCCESS = '[Book Page]Get Book Success',
  OPEN_POPUP = '[Book Page]Open Popup',
}